+++ app/src/main/java/com/example/sadhumster/datasource/App.kt
@@ -1,7 +1,9 @@
-package com.example.sadhumster
+package com.example.sadhumster.datasource
 
 import android.app.Application
-import com.example.sadhumster.db.AppDatabase
+import com.example.sadhumster.datasource.db.AppDatabase
+import dagger.hilt.android.HiltAndroidApp
+
 
 class App : Application() {
     override fun onCreate() {
