+++ app/src/main/java/com/example/sadhumster/datasource/db/AppDataBase.kt
@@ -1,11 +1,11 @@
-package com.example.sadhumster.db
+package com.example.sadhumster.datasource.db
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import com.example.sadhumster.model.Joke
-import com.example.sadhumster.model.JokeFromInternet
+import com.example.sadhumster.domain.model.Joke
+import com.example.sadhumster.domain.model.JokeFromInternet
 
 
 @Database(entities = [Joke::class, JokeFromInternet::class], version = 1)
