+++ app/src/main/java/com/example/sadhumster/domain/model/Joke.kt
@@ -1,0 +1,49 @@
+package com.example.sadhumster.domain.model
+
+import android.os.Parcel
+import android.os.Parcelable
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "jokes")
+data class Joke(
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    val category: String,
+    val setup: String,
+    val delivery: String,
+    val from: String
+) : Parcelable {
+    constructor(parcel: Parcel) : this(
+        parcel.readInt(),
+        parcel.readString().toString(),
+        parcel.readString().toString(),
+        parcel.readString().toString(),
+        parcel.readString().toString()
+    )
+
+    override fun writeToParcel(parcel: Parcel, flags: Int) {
+        parcel.writeInt(id)
+        parcel.writeString(category)
+        parcel.writeString(setup)
+        parcel.writeString(delivery)
+        parcel.writeString(from)
+    }
+
+    override fun describeContents(): Int {
+        return 0
+    }
+
+    companion object CREATOR : Parcelable.Creator<Joke> {
+        override fun createFromParcel(parcel: Parcel): Joke {
+            return Joke(parcel)
+        }
+
+        override fun newArray(size: Int): Array<Joke?> {
+            return arrayOfNulls(size)
+        }
+    }
+}
+
+
+
+
