+++ app/src/main/java/com/example/sadhumster/presentation/fragments/FragmentJokesDetails.kt
@@ -1,4 +1,4 @@
-package com.example.sadhumster.fragments
+package com.example.sadhumster.presentation.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,10 +7,8 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.sadhumster.R
 import com.example.sadhumster.databinding.FragmentJokesDetailsBinding
-import com.example.sadhumster.model.Joke
+import com.example.sadhumster.domain.model.Joke
 
-const val key1 = "jokeIndex"
-const val key2 = "jokesList"
 
 class FragmentJokesDetails : Fragment(R.layout.fragment_jokes_details) {
 
@@ -21,7 +19,7 @@ class FragmentJokesDetails : Fragment(R.layout.fragment_jokes_details) {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View{
+    ): View {
         _binding = FragmentJokesDetailsBinding.inflate(inflater, container, false)
         return binding.root
     }
@@ -32,12 +30,22 @@ class FragmentJokesDetails : Fragment(R.layout.fragment_jokes_details) {
             "jokeDetails",
             viewLifecycleOwner
         ) { _, bundle ->
-            val jokesList = bundle.getParcelableArrayList<Joke>(key2)
-            val jokeIndex = bundle.getInt(key1)
+            val jokesList = bundle.getParcelableArrayList<Joke>(JOKE_LIST_KEY)
+            val jokeIndex = bundle.getInt(JOKE_INDEX_KEY)
             jokesList?.get(jokeIndex)?.let { joke ->
+                var from: String = ""
+                when (joke.from) {
+                    "fromInternet" -> from = context?.getString(R.string.from_internet).toString()
+                    "fromFragment" -> from = context?.getString(R.string.from_fragment).toString()
+                }
                 with(binding) {
                     hamster.setImageResource(R.drawable.hamster)
-                    val s = context?.getString(R.string.joke_title, (jokeIndex + 1), joke.category, joke.from)
+                    val s = context?.getString(
+                        R.string.joke_title,
+                        (jokeIndex + 1),
+                        joke.category,
+                        from
+                    )
                     header.text = s
                     question.text = joke.setup
                     answer.text = joke.delivery
@@ -45,6 +53,7 @@ class FragmentJokesDetails : Fragment(R.layout.fragment_jokes_details) {
             }
         }
     }
+
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
