+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".activities.MainActivity">
+    tools:context=".presentation.activities.MainActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/container"
