+++ app/src/main/res/layout/fragment_jokes_details.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.FragmentJokesDetails">
+    tools:context=".presentation.fragments.FragmentJokesDetails">
 
     <ImageView
         android:id="@+id/hamster"
