+++ app/src/main/res/values/strings.xml
@@ -11,7 +11,7 @@
     <string name="joke_title">#%d %s %s</string>
     <string name="error">Try again</string>
     <string name="from_internet">(internet)</string>
-    <string name="from_fragment">(your)</string>"
-    <!-- TODO: Remove or change this placeholder text -->
+    <string name="from_fragment">(your)</string>
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="try_again">Try again later</string>
 </resources>
